valid2

+++ philo/philo.h
@@ -6,7 +6,7 @@
 /*   By: ssamadi <ssamadi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/16 15:29:53 by ssamadi           #+#    #+#             */
-/*   Updated: 2021/06/30 13:53:40 by ssamadi          ###   ########.fr       */
+/*   Updated: 2021/07/01 11:46:47 by ssamadi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,12 +21,12 @@
 
 typedef struct s_all
 {
-	int time_to_eat;
-	int time_to_sleep;
-	int time_to_die;
-	int number_philo;
-	int	number_philo_must_eat;
-	int last_eat;
+	long long time_to_eat;
+	long long time_to_sleep;
+	long long time_to_die;
+	long long number_philo;
+	long long	number_philo_must_eat;
+	long long last_eat;
 	pthread_mutex_t *forks;
 	pthread_mutex_t write_lock;
 }   t_all;
